# Crossword repair loop fixes