- [x] Fix #113
- [x] Fix #109

Before:
```
goos: linux
goarch: amd64
pkg: github.com/celestiaorg/rsmt2d
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkRepair/Repairing_16x16_ODS_using_RSGF8-16         	      33	  36366217 ns/op
BenchmarkRepair/Repairing_32x32_ODS_using_RSGF8-16         	       4	 291625597 ns/op
BenchmarkRepair/Repairing_64x64_ODS_using_RSGF8-16         	       1	2196795856 ns/op
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-16       	       1	18349927611 ns/op
```

After:
```
goos: linux
goarch: amd64
pkg: github.com/celestiaorg/rsmt2d
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkRepair/Repairing_16x16_ODS_using_RSGF8-16         	     363	   3733920 ns/op
BenchmarkRepair/Repairing_32x32_ODS_using_RSGF8-16         	      82	  20524141 ns/op
BenchmarkRepair/Repairing_64x64_ODS_using_RSGF8-16         	      16	  64611698 ns/op
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-16       	       4	 268046116 ns/op
```