@Wondertan 2022-09-14
@musalbas Can we write a test asserting that #113 is being fixed or make an issue that needs to be tested? It is indeed fixed, and a unit test to concrete this would not hurt.